Register - Login System
This Project involves concept of File Handling .It shows an interface to user for registering or logging in . Every time a user registers, it first checks if a same username already exists ,if a username already exists it notifies user and asks to choose another username otherwise it registers the user and creates a file containing the details of the user. Whenever a user wants to Login ,after entering the Login details it checks if such user is already registered ,if registered then checks the details and logs-in him/her otherwise notifies that the credentials are incorrect.