Casino Betting Game
This game involves use of libraries like <srand> and <ctime> . User is asked to enter the initial starting balance he/she wants to bet with. Then every time he is asked to enter the amount he wants to bet and is asked to guess a number between 0-9. If the number matches then the amount is doubled and added to the balance ,otherwise that amount is deducted from the balance.