Casino Betting Game
This projects involves implementation of C++ libraries like <srand> and <ctime>. User is first asked to enter the balance he wants to start with. Every time the user is asked to enter an amount and guess a number between 0 to 9. If the number matches the random generated number then the amount is doubled and credited to the balance . If the number generated doesn't match then that amount is deducted from his balance.